Dataset center

Generic function to calculate the center of a compositional dataset

center(X, zero.rm = FALSE, na.rm = FALSE)

Arguments

X

compositional dataset

zero.rm

a logical value indicating whether zero values should be stripped before the computation proceeds.

na.rm

a logical value indicating whether NA values should be stripped before the computation proceeds.

Examples

X = matrix(exp(rnorm(5*100)), nrow=100, ncol=5)
g = rep(c('a','b','c','d'), 25)
center(X)
#> [1] 0.1869451 0.2499482 0.1622342 0.1932624 0.2076102
(by_g <- by(X, g, center))
#> INDICES: a
#>        V1        V2        V3        V4        V5 
#> 0.1666041 0.2577717 0.1846411 0.1898717 0.2011114 
#> ------------------------------------------------------------ 
#> INDICES: b
#>        V1        V2        V3        V4        V5 
#> 0.2269412 0.2141864 0.1617915 0.2088780 0.1882029 
#> ------------------------------------------------------------ 
#> INDICES: c
#>        V1        V2        V3        V4        V5 
#> 0.2462371 0.2566625 0.1096549 0.1976456 0.1897999 
#> ------------------------------------------------------------ 
#> INDICES: d
#>        V1        V2        V3        V4        V5 
#> 0.1243905 0.2611523 0.2005123 0.1687458 0.2451992 
center(t(simplify2array(by_g)))
#>        V1        V2        V3        V4        V5 
#> 0.1869451 0.2499482 0.1622342 0.1932624 0.2076102