Get coordinates from compositions w.r.t. an specific basis

Calculate the coordinates of a composition with respect a given basis

coordinates(X, basis = "ilr", label = NULL, sparse_basis = FALSE)

Arguments

X
compositional dataset. Either a matrix, a data.frame or a vector

basis
basis used to calculate the coordinates. basis can be either a string or a matrix. Accepted values for strings are: 'ilr' (default), 'clr', 'alr', 'pc', 'pb' and 'cdp'. If basis is a matrix, it is expected to have log-ratio basis given in columns.

label
name given to the coordinates

sparse_basis
Is the given matrix basis sparse? If TRUE calculation are carried taking into an account sparsity (default `FALSE`)

Value

Coordinates of composition X with respect the given basis.

Details

coordinates function calculates the coordinates of a compositiona w.r.t. a given basis. `basis` parameter is used to set the basis, it can be either a matrix defining the log-contrasts in columns or a string defining some well-known log-contrast: 'alr' 'clr', 'ilr', 'pc', 'pb' and 'cdp', for the additive log-ratio, centered log-ratio, isometric log-ratio, clr principal components, clr principal balances or default's CoDaPack balances respectively.

See also

Examples

coordinates(c(1,2,3,4,5))

#>       ilr1       ilr2       ilr3       ilr4 
#> -0.4901291 -0.6140370 -0.6833297 -0.7288906 

# basis is shown if 'coda.base.basis' option is set to TRUE
options('coda.base.basis' = TRUE)
coordinates(c(1,2,3,4,5))

#>       ilr1       ilr2       ilr3       ilr4 
#> -0.4901291 -0.6140370 -0.6833297 -0.7288906 
#> attr(,"basis")
#>            [,1]       [,2]       [,3]       [,4]
#> [1,]  0.7071068  0.4082483  0.2886751  0.2236068
#> [2,] -0.7071068  0.4082483  0.2886751  0.2236068
#> [3,]  0.0000000 -0.8164966  0.2886751  0.2236068
#> [4,]  0.0000000  0.0000000 -0.8660254  0.2236068
#> [5,]  0.0000000  0.0000000  0.0000000 -0.8944272

# Setting sparse_basi to TRUE can improve performance if log-ratio basis is sparse.
N = 100
K = 1000
X = matrix(exp(rnorm(N*K)), nrow=N, ncol=K)
system.time(coordinates(X, alr_basis(K), sparse_basis = FALSE))

#>    user  system elapsed 
#>   0.091   0.010   0.101 

system.time(coordinates(X, alr_basis(K), sparse_basis = TRUE))

#>    user  system elapsed 
#>   0.024   0.003   0.027 

system.time(coordinates(X, 'alr'))

#>    user  system elapsed 
#>   0.015   0.003   0.018